+++ docs/source/release-notes.rst
@@ -11,8 +11,10 @@ v1.5 Notes
 
 v1.5.2 (2021-10-27)
 ```````````````````
-- Bug fixes.
+- Added ``andes prep -c`` to precompile the generated code.
+- Compiled numba code is always cached.
 - Removed ``__zeros`` and ``__ones`` as they are no longer needed.
+- Bug fixes.
 
 v1.5.1 (2021-10-23)
 ```````````````````