+++ docs/source/release-notes.rst
@@ -11,11 +11,14 @@ v1.5 Notes
 
 v1.5.2 (2021-10-27)
 ```````````````````
-- Added ``andes prep -c`` to precompile the generated code.
-- Compiled numba code is always cached.
+- Removed ``CVXOPT`` dependency.
 - Removed ``__zeros`` and ``__ones`` as they are no longer needed.
+
+- Added ``andes prep -c`` to precompile the generated code.
 - Added utility functions for saving and loading system snapshots.
   See ``andes/utils/snapshot.py``.
+
+- Compiled numba code is always cached.
 - Bug fixes.
 
 v1.5.1 (2021-10-23)